+++ scripts/Groovy/kafka/KafkaQuery.groovy
@@ -1,6 +1,7 @@
 @GrabResolver(root = 'http://packages.confluent.io/maven/', name = 'Confluent')
+@GrabConfig(systemClassLoader = true)
 @Grab(group = 'org.apache.kafka', module = 'kafka-clients', version = '2.3.1')
-@Grab(group = 'ch.qos.logback', module = 'logback-classic', version = '1.2.11')
+@Grab(group = 'ch.qos.logback', module = 'logback-classic', version = '1.4.7')
 @Grab(group = 'io.confluent', module = 'kafka-avro-serializer', version = '5.2.2')
 @Grab(group = 'com.jayway.jsonpath', module = 'json-path', version = '2.4.0')
 @Grab(group = 'io.micrometer', module = 'micrometer-registry-jmx', version = '1.5.7')